requirements.txt
Django
django-allauth
django-simple-captcha
django-ckeditor
pillow
django-notifications-hq
django-filter